requirements.txt
amqp==2.5.0
backcall==0.1.0
billiard==3.6.0.0
boto3==1.9.163
botocore==1.12.163
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-channels==0.7.0
django-debug-toolbar==1.11
django-jsonfield==1.2.0
django-multiselectfield==0.1.8
django-rest-auth==0.9.5
django-storages==1.7.1
django-timezone-field==3.0
djangorestframework==3.9.4
docutils==0.14
geographiclib==1.49
geopy==1.20.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.1
mysqlclient==1.4.2.post1
numpy==1.16.4
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyfcm==1.4.7
Pygments==2.4.2
python-crontab==2.3.6
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
raven==6.10.0
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
